refactor: merge Vite plugins (data-inject, build-assets) and remove unused plugins
- Merge manifest-inject + site-config-inject → data-inject
- Merge og-image + feed-sitemap → build-assets
- Delete unused i18n-hmr.ts and locales.ts
- Update vite.config.ts imports
